refactor: move shell hooks in separated files
