+++ src/boot/theme-provider.tsx
@@ -22,10 +22,10 @@ import { reduce } from 'lodash';
 import { createGlobalStyle, DefaultTheme } from 'styled-components';
 import { DarkReaderPropValues, ThemeExtension } from '../../types';
 import { darkReaderDynamicThemeFixes, LOCAL_STORAGE_SETTINGS_KEY } from '../constants';
-import { useLocalStorage } from '../shell/hooks';
 import { ScalingSettings } from '../../types/settings';
 import { getAutoScalingFontSize } from '../settings/components/utils';
 import { useGetPrimaryColor } from './use-get-primary-color';
+import { useLocalStorage } from '../shell/hooks/useLocalStorage';
 
 setFetchMethod(window.fetch);
