+++ src/shell/hooks/useIsMobile.ts
@@ -1,0 +1,12 @@
+/*
+ * SPDX-FileCopyrightText: 2023 Zextras <https://www.zextras.com>
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only
+ */
+import { useContext } from 'react';
+import ShellContext from '../shell-context';
+
+export function useIsMobile(): boolean {
+	const { isMobile } = useContext(ShellContext);
+	return isMobile;
+}
